+++ src/tests/performance/snr/mod.rs
@@ -1,1 +1,2 @@
 pub mod ft8;
+pub mod psk31;